+++ Source/Core/VideoCommon/G_GZLE01_pvt.h
@@ -4,5 +4,5 @@
 class G_GZLE01_pvt
 {
 public:
-static void Initialize(std::map<size_t, TCompiledLoaderFunction> &pvlmap);
+static void Initialize(std::map<u64, TCompiledLoaderFunction> &pvlmap);
 };
